Introduced the function to convert the concentration of calcium in number of particles.
The volume is imposed by a default parameter but it should be taken by the spine head automatically. This will give the automatic precision if we change the volume of the spine.

The spine's volume in ecell and in NEURON should be the same and the one in ecell should depend on NEURON one.