I removed most of the jars neede by the old Biopax converter, as paxtools should include everything it needs.
I cleaned and corrected the generalConverter.sh script to be able to use it for several converter and on the biomodels folders.

I removed the LibSBMLUtilities class to be sure to have no dependency to libSBML. Then I replace all the call in SBML2XPP and SBML2Octave by the proper jsbml call.
I have done some minors corrections on the SBML2XPP