Fixed ignoring of primary flag for resources.
Replaced usage of Groovy each with for-each loop to ensure that when
a primary resource is found we return it and exit. calling return from
a closure breaks from the closure, but not the method.

Reworked the resolver to use a TreeSet with a custom comparator rather
than a HashSet to keep track of a data collection's resources. This avoids
the need to process the resources twice -- once to figure out the max uptime
and once again to get the id.

Finally, worked around a nasty bug whereby accessing a collection's resources
through its resources property would only yield one of its resources rather than
all of them. This could be a result of either the usage of static methods in the
Resolver service, or of an issue with the association fetching strategy (FM.JOIN).
Regardless, calling DataCollection.findBy().getResources() is slower than
Resource.findAllByDataCollection(dataCollectionProxy), so it's not a bad change to
make.